refactor(extract_word): move extract_inhibited_content to extract_word.c and improve memory management in safe_strcat
